refactor: share _exit_reaching_nodes for the straight-mode trunk fix (#1426)
Hoist _exit_reaching_nodes into _common so _materialize_pass_through_lines
reuses it instead of an inline copy of the same reverse-BFS; ports.py and
single_section.py now import it from _common. Fold the straight-mode
full-bundle-dead-end competitor case into the function docstring and drop the
duplicated inline comment.
