+++ src/nf_metro/layout/phases/_common.py
@@ -1200,6 +1200,29 @@ def _section_fan_trunk_lines(graph: MetroGraph, section: Section) -> set[str]:
     return trunk or bundle
 
 
+def _exit_reaching_nodes(graph: MetroGraph, section: Section) -> frozenset[str]:
+    """Internal stations with a forward path to one of *section*'s exit ports.
+
+    Walks backward from each exit port through in-section, non-port feeders, so
+    a station is included when the flow through it continues out of the section
+    (the section's through-line) rather than dead-ending inside it.  A terminal
+    section with no exit port yields the empty set.
+    """
+    sec_ids = set(section.station_ids)
+    reaching: set[str] = set()
+    stack = list(section.exit_ports)
+    while stack:
+        node = stack.pop()
+        for edge in graph.edges_to(node):
+            src = edge.source
+            st = graph.stations.get(src)
+            if src in reaching or src not in sec_ids or st is None or st.is_port:
+                continue
+            reaching.add(src)
+            stack.append(src)
+    return frozenset(reaching)
+
+
 def _section_row_through_lines(graph: MetroGraph, section: Section) -> set[str]:
     """Lines on a section's LEFT/RIGHT ports that connect to a same-row section.
